# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.CreateAdminRequest:
    properties:
      email:
        example: admin@example.com
        type: string
      password:
        example: password123
        type: string
      username:
        example: admin
        type: string
    type: object
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.CreateAdminResponse:
    properties:
      accessToken:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      refreshToken:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_models.User'
    type: object
...
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.CreateUserPayload:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_models.Userroles'
        enum:
        - Admin
        - Manager
        - Member
      teamId:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - role
    - username
    type: object
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.CreateUserResponse:
    properties:
      user:
        $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_models.User'
    type: object
...
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.MessageResponse:
    properties:
      message:
        example: operation completed successfully
        type: string
    type: object
...
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.UpdateUserPayload:
    properties:
      email:
        type: string
      username:
        type: string
    required:
    - email
    - username
    type: object
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.UserExistsResponse:
    properties:
      exists:
        example: true
        type: boolean
    type: object
  github_com_TobiasRV_challenge-fs-senior_internals_interfaces.UsersListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_models.User'
        type: array
      pagination:
        $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.Pagination'
    type: object
...
  github_com_TobiasRV_challenge-fs-senior_internals_models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      role:
        $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_models.Userroles'
      teamId:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
  github_com_TobiasRV_challenge-fs-senior_internals_models.Userroles:
    enum:
    - Admin
    - Manager
    - Member
    type: string
    x-enum-varnames:
    - UserrolesAdmin
    - UserrolesManager
    - UserrolesMember
  github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorBody:
    properties:
      body:
        example: error message
        type: string
    type: object
  github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse:
    properties:
      errors:
        $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorBody'
    type: object
  github_com_TobiasRV_challenge-fs-senior_internals_utils.Pagination:
    properties:
      next_cursor:
        example: eyJpZCI6IjEyMzQ1In0=
        type: string
      prev_cursor:
        example: eyJpZCI6IjEyMzQ1In0=
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for managing teams, projects, tasks, and users
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Challenge FS Senior API
  version: "1.0"
paths:
...
  /auth/register-admin:
    post:
      consumes:
      - application/json
      description: Create a new admin user account
      parameters:
      - description: Admin registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.CreateAdminRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.CreateAdminResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
      summary: Register admin user
      tags:
      - Auth
...
  /users:
    get:
      consumes:
      - application/json
      description: Get paginated list of users (Admin or Manager only)
      parameters:
      - description: Filter by email
        in: query
        name: email
        type: string
      - description: Team ID (required)
        in: query
        name: teamId
        required: true
        type: string
      - description: Filter by role (Admin, Manager, Member)
        in: query
        name: role
        type: string
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Number of items per page
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.UsersListResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "403":
          description: Forbidden - Admin or Manager only
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user (Admin only)
      parameters:
      - description: User creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.CreateUserPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.CreateUserResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new user
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID (Admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.MessageResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user information (Admin only)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.UpdateUserPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_models.User'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - Users
  /users/exists-by-email:
    get:
      consumes:
      - application/json
      description: Check if a user with the given email exists
      parameters:
      - description: User email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_interfaces.UserExistsResponse'
        "400":
          description: Email is required
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_TobiasRV_challenge-fs-senior_internals_utils.ErrorResponse'
      summary: Check if user exists by email
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"